[S12P21E206-187] 정산하기 기능 오더,셀러 구현

+++ back/funding/src/main/java/com/ssafy/funding/controller/FundingController.java
@@ -152,7 +152,7 @@ public class FundingController {
 
     // 정산 완료 후 Funding 서비스의 settlementCompleted 플래그를 업데이트 (true로 변경)
     @PostMapping("/update-event-sent")
-    public void updateSettlementCompleted(@RequestParam Long fundingId, @RequestParam Boolean eventSent) {
+    public void updateSettlementCompleted(@RequestParam int fundingId, @RequestParam Boolean eventSent) {
         productService.updateSettlementCompleted(fundingId, eventSent);
     }
 }
